Name : Mitul Module : CTMS SubModule : 484 Description : Patient Cost

+++ GSC.Data/Entities/CTMS/Procedure.cs
@@ -12,7 +12,7 @@ namespace GSC.Data.Entities.CTMS
     {
         public string Name { get; set; }    
         public int? UnitId { get; set; }
-        public int? CostPerUnit { get; set; }    
+        public decimal? CostPerUnit { get; set; }    
         public int? CurrencyId { get; set; }
         public Unit Unit { get; set; }
         public Currency Currency { get; set; }
